requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.21.0
pandas>=1.3.0
pickle-mixin>=1.0.2
python-docx>=0.8.11
matplotlib>=3.5.0
plotly>=5.0.0
scikit-learn>=1.0.0